test(scratchpad): move FirstFit determinism test into ScoreOrderingTests
Fold the cross-process PYTHONHASHSEED regression test into the shared
ScoreOrderingTests mixin in test_scratchpad_solver.py, instead of a
standalone file, so BestFitLayoutSolver is covered too (it inherits the
same _build_gaps in-place-parent-selection logic from FirstFitLayoutSolver
that the test guards). Drop the now-unneeded OOT test config entry for
the removed file, and update ScoreOrderingTests' docstring to describe
the added determinism check and why it doesn't apply to Greedy.
