--- tests/configs/torch_spyre_tests/inductor/test_firstfit_determinism_config.yaml
@@ -1,5 +1,0 @@
-test_suite_config:
-  labels: [core, full]
-  files:
-    - path: ${TORCH_DEVICE_ROOT}/tests/inductor/test_firstfit_determinism.py
-      unlisted_test_mode: mandatory_success
